requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai>=1.50.0
gunicorn==23.0.0
python-dotenv==1.0.0